requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.11
fonttools==4.56.0
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
macholib==1.16.3
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
wheel==0.45.1
